Extract advanced config to its own component
